requirements.txt
django-leaflet
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
